Car_Build/P-Api/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.118.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-tools==1.75.1
h11==0.16.0
idna==3.10
protobuf==6.32.1
pydantic==2.11.9
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
